% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_biomass.R
\name{compute_biomass}
\alias{compute_biomass}
\title{Compute the biomass of a tree list}
\usage{
compute_biomass(data, species, diameter, height, total = TRUE)
}
\arguments{
\item{data}{Data frame containing the tree list.}

\item{species}{Character. Name of the column of \code{data} containing the
species information.}

\item{diameter}{Character. Name of the column of \code{data} containing the
diameter information.}

\item{height}{Character. Name of the column of \code{data} containing the
height information.}

\item{total}{Logical. Should the output also contain the total biomass?
Defaults to \code{TRUE}}
}
\value{
Data frame containing the biomass estimates of the trees in
\code{data} for different components (in kilograms), along with all other
columns of \code{data}.
}
\description{
Using allometric equations, this function estimates the biomass of different
tree components using the species, diameter, and (if available) height of the
tree.
}
\details{
Biomass is estimated for the corresponding tree using the equations published
in Lambert \emph{et al} (2005) and Ung \emph{et al} (2008). The estimates are given for
four different components of the tree: bark, branches, foliage, and wood. If
\code{total = TRUE} (which is the default), the output also contains the
total biomass (i.e. adding up all four components).

The user must pay special attention to the units of the input data. Heights
should be given in meters, while diameters (at breast height) should be given
in centimeters.
}
\examples{
test_df <- data.frame(
    'spec' = c("Jack pine", "Trembling aspen", "Black spruce",
               "Black spruce", "Eucaplyptus"),
    'diam' = c(21.0, 32.1, 31.0, 21.0, 14.0),
    'ht' = c(15.2, 28.0, 22.6, 20.8, 10.2)
)
compute_biomass(test_df, species = 'spec', diameter = 'diam',
                height = 'ht')
# Using diameters only
compute_biomass(test_df, species = 'spec', diameter = 'diam')
}
\references{
Lambert, M.-C., C.-H. Ung, and F. Raulier 2005. Canadian national
biomass equations. Can. J. For. Res 35: 1996-2018.

Ung, C.-H., Bernier, P., Guo, X.-J. 2008. Canadian national
biomass equations: new parameter estimates that include British Columbia
data. Can. J. For. Res 38:1123-2232.
}
