% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_biomass.R
\name{compute_biomass_single_tree}
\alias{compute_biomass_single_tree}
\title{Compute the biomass of a single tree}
\usage{
compute_biomass_single_tree(species, diameter, height, wide = FALSE)
}
\arguments{
\item{species}{Species of the tree}

\item{diameter}{Diameter of the tree at breast height (in centimeters)}

\item{height}{Height of the tree (in meters)}

\item{wide}{Logical. Should the output be transformed into wide format?
Defaults to \code{FALSE}}
}
\value{
Data frame containing the biomass estimates of the different
components (in kilograms)
}
\description{
Using allometric equations, this function estimates the biomass of different
tree components using the species, diameter, and (if available) height of the
tree.
}
\details{
Biomass is estimated for the corresponding tree using the equations published
in Lambert \emph{et al} (2005) and Ung \emph{et al} (2008). The estimates are given for
four different components of the tree: bark, branches, foliage, and wood.

The user must pay special attention to the units of the input data. Heights
should be given in meters, while diameters (at breast height) should be given
in centimeters.
}
\examples{
compute_biomass_single_tree(species = 'Jack pine', diameter = 21, height = 15.2)
compute_biomass_single_tree(species = 'Jack pine', diameter = 21)
# Wide format
compute_biomass_single_tree(species = 'Jack pine', diameter = 21, height = 15.2, wide = TRUE)
}
\references{
Lambert, M.-C., C.-H. Ung, and F. Raulier 2005. Canadian national
biomass equations. Can. J. For. Res 35: 1996-2018.

Ung, C.-H., Bernier, P., Guo, X.-J. 2008. Canadian national
biomass equations: new parameter estimates that include British Columbia
data. Can. J. For. Res 38:1123-2232.
}
