% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_species}
\alias{show_species}
\title{Show the species name available}
\usage{
show_species(dataset = c("diameter-height", "diameter-only", "plot-level"))
}
\arguments{
\item{dataset}{From which dataset should we pull the names?}
}
\value{
Character string with the unique species name.
}
\description{
Show the species name available in each of the three databases of allometric
equations.
}
